refactor(portal): extract asset bundle helpers
